feat: card project components was created
