+++ data/routes.ts
@@ -1,27 +1,27 @@
-import {Page} from '~/types/index';
+import {Page} from '../interfaces/index';
 
 export const pages: Page[] = [
-	{
-		name: 'Home',
-		url: '/'
-	},
-	{
-		name: 'Projects',
-		url: '/projects'
-	},
-	{
-		name: 'Contact',
-		url: '/contact'
-	},
-	{
-		name: 'Blog',
-		url: '/blog'
-	}
+  {
+    name: 'Home',
+    url: '/',
+  },
+  {
+    name: 'Projects',
+    url: '/projects',
+  },
+  {
+    name: 'Contact',
+    url: '/contact',
+  },
+  {
+    name: 'Blog',
+    url: '/blog',
+  },
 ];
 
 export const settings: Page[] = [
-	{
-		name: 'Download CV',
-		url: '/download-cv'
-	}
+  {
+    name: 'Download CV',
+    url: '/download-cv',
+  },
 ];
