+++ interfaces/index.ts
@@ -1,0 +1,13 @@
+export interface Page {
+  name: string,
+  url: string
+}
+
+export interface Projects {
+  name: string,
+  image: string,
+  description: string,
+  urlGitHub: string,
+  urlPreview: string,
+  stack: string[]
+}
