+++ tailwind.config.ts
@@ -21,6 +21,6 @@ const config: Config = {
       'xl': '1280px',
     },
   },
-  plugins: [require("tailwindcss-animate")],
+  plugins: [require('tailwindcss-animate')],
 };
 export default config;
