--- types/index.ts
@@ -1,4 +1,0 @@
-export interface Page {
-	name: string,
-	url: string
-}
